harvardcards/requirements/heroku.txt
-r base.txt

django-toolbelt==0.0.1
static3==0.5.1
wsgiref==0.1.2
